Fixes #85 - iOS header icons too dark in dark mode
Screen capture of new behaviour can be found in Discord -
https://discord.com/channels/677546901339504640/1060546605008298075/1060546670686916659

Looks like Safari on iOS either didn't like the invert filter, or didn't
like the selector we used to target the SVG path elements for inverting
when dark mode is enabled. I don't have an OSX machine so can't inspect
the browser state to say for sure, so that's just a guess.

Regardless of the cause, this commit puts a workaround in place that
uses a React media query hook to change the content of the SVG returned
by the image components directly, changing their colour depending
whether dark mode is enabled or not. This necessitated a new package in
the project, namely `react-media-hook`

I also had to slightly change some of the adjecent CSS to remove the
now-redundant inverting filter on these path elements, and due to
the selector now becoming unreliable with the new changing CSS I elected
to just swap the rules that matched on CSS properties for ones that
matches a selector class. Feels weird to use a class just to _un_target
some behaviour but is what it is.