Move fix to source files for generated icons
Moves the light / dark styling of the SVG icons into CSS and undoes all
changes to generated icon files. Behaviour is now controlled by a class
present in the source that the icon builder allows to exist in the
generated output, so changes will now persist even when icons are
regenerated.